+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.5 — 2025-12-16
+
+- Implemented central FD method for second-order precision, set to `--method=central` in both `zstar gen` and `zstar deal` to run it, defalut still set as `--method=forward` to save computing resources.
+
+## 0.0.4 — 2025-12-12
+
+- Remove `out_chg 1 10` style, just use `out_chg 1`.
+  
 ## 0.0.3 — 2025-12-11
 
 - Fix bugs in the post-processing of Born effective charges (BEC) for the ABACUS NSCF backend, including symmetry reconstruction and automatic generation of `Z-BORN-symm.out`.